+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/annotation/annotationusestyle/InputAnnotationUseStyleCustomAnnotation.java
@@ -1,0 +1,4 @@
+package com.puppycrawl.tools.checkstyle.checks.annotation.annotationusestyle;
+
+@interface InputAnnotationUseStyleCustomAnnotation {
+}
